+++ .gitignore
@@ -61,3 +61,6 @@ buck-out/
 # Expo
 .expo/*
 web-build/
+
+#Firebase
+.firebase
